% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_describe_backups}
\alias{cloudhsmv2_describe_backups}
\title{Gets information about backups of AWS CloudHSM clusters}
\usage{
cloudhsmv2_describe_backups(NextToken, MaxResults, Filters,
  SortAscending)
}
\arguments{
\item{NextToken}{The \code{NextToken} value that you received in the previous response. Use
this value to get more backups.}

\item{MaxResults}{The maximum number of backups to return in the response. When there are
more backups than the number you specify, the response contains a
\code{NextToken} value.}

\item{Filters}{One or more filters to limit the items returned in the response.

Use the \code{backupIds} filter to return only the specified backups. Specify
backups by their backup identifier (ID).

Use the \code{sourceBackupIds} filter to return only the backups created from
a source backup. The \code{sourceBackupID} of a source backup is returned by
the CopyBackupToRegion operation.

Use the \code{clusterIds} filter to return only the backups for the specified
clusters. Specify clusters by their cluster identifier (ID).

Use the \code{states} filter to return only backups that match the specified
state.}

\item{SortAscending}{}
}
\description{
Gets information about backups of AWS CloudHSM clusters.
}
\details{
This is a paginated operation, which means that each response might
contain only a subset of all the backups. When the response contains
only a subset of backups, it includes a \code{NextToken} value. Use this
value in a subsequent \code{DescribeBackups} request to get more backups.
When you receive a response with no \code{NextToken} (or an empty or null
value), that means there are no more backups to get.
}
\section{Request syntax}{
\preformatted{cloudhsmv2$describe_backups(
  NextToken = "string",
  MaxResults = 123,
  Filters = list(
    list(
      "string"
    )
  ),
  SortAscending = TRUE|FALSE
)
}
}

\keyword{internal}
